found community, roadmap, ~team, FAQ, github, footer
